pyproject.toml
[project]
name = "service-mcp-server"
version = "0.1.0"
description = "MCP Server for your OpenRemote instance."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp==2.13.0.2",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.38.0",
    "openremote-client==1.1.3",
    "jinja2>=3.1.6",
]
